+++ CHANGELOG.md
@@ -4,6 +4,8 @@ As this project is pre 1.0, breaking changes may happen for minor version bumps.
 
 ## Next version
 
+- All generated transaction are v1 in support of protocol 13.
+
 ## 0.12.2
 
 As requested by SDF, support for multiplexed account addressing has been removed. This feature has not yet